Some changes dunno what

+++ wypozyczalnia.Server/DTOs/VehicleDTO.cs
@@ -9,11 +9,11 @@ namespace search.api.DTOs
         public string Model { get; set; } = string.Empty; 
         public int YearOfProduction { get; set; } 
         public string Type { get; set; } = string.Empty;
-        public decimal Price { get; set; } 
+        public float Price { get; set; } 
         public string DriveType { get; set; } = string.Empty;
         public string Transmission { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty; 
-        public decimal Rate { get; set; }
+        public int Rate { get; set; }
         public string Localization { get; set; } = string.Empty;
         public string SerialNo { get; set; } = string.Empty;
         public string Vin { get; set; } = string.Empty;
